Correctly handle ResTable_typeSpec by using unsigned bytes to prevent overflow
 - fixes #1185
 - allows applications that have large typeSpec indexes to work

+++ brut.apktool/apktool-lib/src/main/java/brut/androlib/res/data/ResTypeSpec.java
@@ -30,10 +30,10 @@ public final class ResTypeSpec {
     private final ResTable mResTable;
     private final ResPackage mPackage;
 
-    private final byte mId;
+    private final int mId;
     private final int mEntryCount;
 
-    public ResTypeSpec(String name, ResTable resTable, ResPackage package_, byte id, int entryCount) {
+    public ResTypeSpec(String name, ResTable resTable, ResPackage package_, int id, int entryCount) {
         this.mName = name;
         this.mResTable = resTable;
         this.mPackage = package_;
@@ -45,7 +45,7 @@ public final class ResTypeSpec {
         return mName;
     }
 
-    public byte getId() {
+    public int getId() {
         return mId;
     }
